ci: split facet cypress tests
https://coveord.atlassian.net/browse/KIT-3445
